% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.infos.grids.hbm.basic.R
\name{get.infos.grids.hbm.basic}
\alias{get.infos.grids.hbm.basic}
\title{Get basic grid data of the BSH-setup for HBM}
\usage{
get.infos.grids.hbm.basic()
}
\value{
list with grid information (see yourself)
}
\description{
Get basic grid data of the BSH-setup for HBM
}
\examples{

  # get grid information
  grid_info <- get.infos.grids.hbm.basic()
  
}
\author{
Daniel Neumann, daniel.neumann@io-warnemuende.de
}
