% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.varmapping.hbm.netcdf.R
\name{get.varmapping.hbm.netcdf}
\alias{get.varmapping.hbm.netcdf}
\title{Get mapping of "netCDF variables names for HBM river input files" to short names}
\usage{
get.varmapping.hbm.netcdf()
}
\value{
nested lists; look at the output yourself
}
\description{
This mapping maps netCDF variable names as used in HBM river input files
to corresponding short names, which are often used.
}
\examples{

  # get variable mapping
  varmapping <- get.varmapping.hbm.netcdf()
  
}
\seealso{
write.river.netCDF
}
\author{
Daniel Neumann, daniel.neumann@io-warnemuende.de
}
