% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.river.mom.R
\name{interpolate.river.mom.R}
\alias{interpolate.river.mom.R}
\title{Title}
\usage{
interpolate.river.mom.R(inData, from = "none", to = "none",
  method = "step")
}
\arguments{
\item{inData}{list containing a data.frame (in$data) with river inflow data, a character array (in$units) with corresponding units, and a character (in$type).}

\item{from}{character, source time interval; e.g. 'annual', 'monthly', or 'daily'}

\item{to}{character, target time interval; e.g. 'annual', 'monthly', or 'daily'}

\item{method}{character, interpolation method}
}
\value{
list of newly averaged data; same structure as input data
}
\description{
Title
}
\examples{

  TODO
}
