% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.river.mom.R
\name{interpolate.river.mom}
\alias{interpolate.river.mom}
\title{Interpolates river inflow data}
\usage{
interpolate.river.mom(inData, from = inData$tstep, to = "none",
  method = "step")
}
\arguments{
\item{inData}{list containing a data.frame (in$data) with river inflow data, a character array (in$units) with corresponding units, and a character (in$format).}

\item{from}{character: current data averaging interval; does not need to be set if in$tstep is set properly; if not, it can be manually set to e.g. 'annual', 'monthly', or 'daily'}

\item{to}{character, target time interval; e.g. 'annual', 'monthly', or 'daily'}

\item{method}{character, interpolation method}
}
\value{
list of newly averaged data; same structure as input data

          list containing a data.frame (out$data), a character array 
          (out$units), and two character variables (out$format and
          out$tstep). The first contains the actual data formatted as a 
          data.frame. The second contains the units to the corresponding 
          columns of the data.frame. The third contains the source/format 
          of data (here: 'mom'; can also be 'swat'). The fourth contains
          information on the time step of the data (resp.: on which time
          interval they are averaged).
}
\description{
Interpolates river inflow data in the time dimensions. Currently, only
interpolation from monthly to values is implemented. Additionally, only 
the interpolation methods 'step' is implemented.
}
\examples{

  # read a file:
  test.mom.monthly <- read.mom('files/GER_Dan_Str_Warnow.dat')
  
  # interpolate from monthly to daily
  test.mom.daily <- interpolate.river.mom(test.mom.monthly, to = 'daily', method = 'step')
  
}
\seealso{
read.mom, write.river.netCDF, write.river.append2Inflow, write.river.newInflow, mean.river.mom
}
\author{
Daniel Neumann, daniel.neumann@io-warnemuende.de
}
