% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.leapyear.R
\name{is.leapyear}
\alias{is.leapyear}
\title{Is a year a leap year?}
\usage{
is.leapyear(year)
}
\arguments{
\item{year}{a numeric giving a year}
}
\value{
Returns a logical. It is TRUE when 'year' is a leap year according to the Gregorian calender. Else it is false.
}
\description{
Provides information (TRUE/FALSE) whether a year is a leap year (TRUE) or not (FALSE).
}
\examples{
  is.leapyear(2012)
  # return: TRUE
  
  is.leapyear(2001)
  # return: FALSE
  
  is.leapyear(1998:2010)
  # return: FALSE FALSE  TRUE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE  TRUE FALSE FALSE
}
\author{
Daniel Neumann, daniel.neumann@io-warnemuende.de
}
