% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.river.mom.R
\name{mean.river.mom}
\alias{mean.river.mom}
\title{Calculates annual mean values from MOM river input data}
\usage{
\method{mean}{river.mom}(inData, from = "none", to = "none")
}
\arguments{
\item{inData}{list containing a data.frame (in$data) with river inflow data, a character array (in$units) with corresponding units, and a character (in$type).}

\item{from}{character: current data averaging interval; e.g. 'annual', 'monthly', or 'daily'}

\item{to}{character: future averaging interval; e.g. 'annual', 'monthly', or 'daily'}
}
\value{
list of newly averaged data; same structure as input data
}
\description{
TODO
}
\examples{

  TODO
}
