% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.infos.rivers.R
\name{read.infos.rivers}
\alias{read.infos.rivers}
\title{Read csv file with information on rivers and return it in a nested list}
\usage{
read.infos.rivers(filename, grids)
}
\arguments{
\item{filename}{path and name of the csv file with river information}

\item{grids}{nested list with grid definitions: grids$fine and grids$coarse are grid definitions of the fine and coarse, respectively BSH-HBM grids}
}
\value{
nested list with grid information it contains:
}
\description{
Columns in the csv file: id, name, lat, lon, grid, ycell, xcell, filename
}
\details{
Elements of the output list: filename, lon, lat, xcell, ycell, grid, riverid, rivername
}
\examples{
  
  # get grid info:
  grid_info <- get.infos.grids.hbm.basic()
  
  # set filename
  file <- 'files/river_list.dat'
  
  # get river infos
  riverInfos <- read.infos.rivers(file, grid_info)
  # (you will get some warnings here)
  
Elements of the output list:
  filename
  lon
  lat
  xcell
  ycell
  grid [fine,coarse]
  riverid
  rivername
}
\seealso{
get.infos.grids.hbm.basic
}
\author{
Daniel Neumann, daniel.neumann@io-warnemuende.de
}
