% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.river.mom.R
\name{read.river.mom}
\alias{read.river.mom}
\title{Reads a river runoff/nutrient-load input file for the ocean model MOM into R}
\usage{
read.river.mom(filename)
}
\arguments{
\item{filename}{character: path/filename of the file to read}
}
\value{
list containing a data.frame (out$data), a character array 
          (out$units), and two character variables (out$format and
          out$tstep). The first contains the actual data formatted as a 
          data.frame. The second contains the units to the corresponding 
          columns of the data.frame. The third contains the source/format 
          of data (here: 'mom'; can also be 'swat'). The fourth contains
          information on the time step of the data (resp.: on which time
          interval they are averaged). The latter is set to 'monthly'.
}
\description{
Reads a MOM input file, which contains river runoff and riverine nutrient 
loads, and writes the data into a data.frame nested into a list. The list
also contains units corresponding to the columns of the data.frame.
}
\examples{

  # read a file:
  test.mom <- read.river.mom('files/GER_Dan_Str_Warnow.dat')
  
}
\seealso{
read.swat, read.swat.data.rch.2012
}
\author{
Daniel Neumann, daniel.neumann@io-warnemuende.de
}
