% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.swat.R
\name{read.swat}
\alias{read.swat}
\title{Reads in a SWAT output file}
\usage{
read.swat(filename, filetype = substr(filename, nchar(filename) - 2,
  nchar(filename)), version = 2016, variables = NULL)
}
\arguments{
\item{filename}{character: name of the SWAT file}

\item{filetype}{character: official file extension of SWAT output file ('rch', 'sub', 'hru')}

\item{version}{integer/numeric: SWAT output version, only 2012 and 2016 (equal) are implemented}

\item{variables}{character array of variable names (==columns) to read in from the file; if == NULL [default], all variables are read}
}
\value{
list containing a data.frame (out$data), a character array 
          (out$units), and a character (out$type). The first contains the 
          actual data formatted as a data.frame (more details below). The 
          second contains the units to the corresponding columns of the 
          data.frame. The third contains the type of data (here: 'swat'; 
          can also be 'mom').
          
          The data.frame out$data which contains the data from the file 
          'filename'; column names of as.data.frame should be equal to 
          column names in the reach file; only variables/columns listed 
          in 'variables' are returned;
}
\description{
Reads a SWAT output file of the *.rch (reach), *.sub, and *.hru formats. 
Currently, only the formats as documented in the SWAT 2012 handbook are 
supported.
}
\examples{
  
  # read in a swat *.rch (reach) file
  data.swat <- read.swat('output.rch')
  
  # read in a swat *.sub file
  data.swat <- read.swat('output.sub')
  # NOTE: not implemented yet
  
  # read in a swat reach file with another file extension
  data.swat <- read.swat('output.txt', filetype = 'rch')
  
  # read in specific variables of a swat *.rch (reach) file
  data.swat <- read.swat('output.rch', variables = c('NO3_OUT', 'NH4_OUT'))
  # NOTE: not implemented yet
  
}
