% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.river.append2Inflow.R
\name{write.river.append2Inflow}
\alias{write.river.append2Inflow}
\title{Appends new inflow information to an existing HBM inflow file}
\usage{
write.river.append2Inflow(riverNames, riverInfos, riverData, grids, dIn, dOt,
  year, month = 0, day = 0, overwrite = TRUE, warn = TRUE)
}
\arguments{
\item{riverNames}{character array with river names}

\item{riverInfos}{list() of grid-information-lists(); to each element in riverNames one list element with the same name needs to exist in riverInfos}

\item{riverData}{list() of river-inflow-data-lists(); to each element in riverData one list element with the same name needs to exist in riverInfos}

\item{grids}{character array with grid names}

\item{dIn}{character; directory containing the input file (to which data should be appended)}

\item{dOt}{character; directory into which the final file should be written}

\item{year}{integer; year to write out}

\item{month}{integer; month to write out; if month==0 => write all months}

\item{day}{integer; day to write out; if day==0 => write all days}

\item{overwrite}{logical; overwrite 'dOt/filename' it already exists}

\item{warn}{logical; allow/suppress warnings}
}
\value{
file written
}
\description{
New river inflow data of the the Fortran format 'F8.6' 
(width=8; 6 digits in front of the '.') is appended to an existing file.
One file per day is modified.
}
\examples{

  # read a file:
  test.mom.monthly <- list()
  test.mom.monthly$Warnow <- read.mom('files/GER_Dan_Str_Warnow.dat')
  test.mom.monthly$Trave <- read.mom('files/GER_Dan_Str_Trave.dat')
  
  # calculate annual means from monthly data
  test.mom.annual <- list()
  test.mom.daily$Warnow <- interpolate.river.mom(test.mom.monthly$Warnow, to = 'daily', method = 'step')
  test.mom.daily$Trave <- interpolate.river.mom(test.mom.monthly$Trave, to = 'daily', method = 'step')
  
  # get grid info
  grid_info <- get.infos.grids.hbm.basic()
  
  # get river infos
  file <- 'files/river_list.dat'
  riverInfos <- read.infos.rivers(file, grid_info)
  
  # write new namelist
  write.river.append2Inflow(c('Warnow', 'Trave'), riverInfos, test.mom.daily, 'files', 'out_dir', 2012, month = 1, day = 1:5, overwrite=FALSE)
  
}
\seealso{
write.river.newInflow, write.river.append2NML
}
\author{
Daniel Neumann, daniel.neumann@io-warnemuende.de
}
