% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.river.netCDF.R
\name{write.river.netCDF}
\alias{write.river.netCDF}
\title{Write netCDF files (one for each river) with riverine nutrient input data}
\usage{
write.river.netCDF(riverNames, riverInfo, riverData, netCDF2input, dOt, year,
  overwrite = TRUE, warn = TRUE, cmd_ncgen = "ncgen")
}
\arguments{
\item{riverNames}{character array with river names}

\item{riverInfo}{list() of grid-information-lists(); to each element in riverNames one list element with the same name needs to exist in riverInfos}

\item{riverData}{list() of river-inflow-data-lists(); to each element in riverData one list element with the same name needs to exist in riverInfos}

\item{netCDF2input}{list() with two character arrays netCDF2input$untagged and netCDF2input$tagged; these contain the names of rivers, which are tagged and untagged, respectively}

\item{dOt}{character; directory into which the final file should be written}

\item{year}{integer; year to write out}

\item{overwrite}{logical; overwrite 'dOt/filename' it already exists}

\item{warn}{logical; allow/suppress warnings}

\item{cmd_ncgen}{character: ncgen-call; if ncgen is not in the path (or the wrong verson of ncgen) please provide here '/path/ncgen-exec' for the ncgen binary}
}
\description{
For each entry in 'riverNames' a netCDF file is created. TODO
}
\examples{

  # read a file:
  test.mom.monthly <- list()
  test.mom.monthly$Warnow <- read.river.mom('files/GER_Dan_Str_Warnow.dat')
  test.mom.monthly$Trave <- read.river.mom('files/GER_Dan_Str_Trave.dat')
  
  # calculate annual means from monthly data
  test.mom.annual <- list()
  test.mom.daily$Warnow <- interpolate.river.mom(test.mom.monthly$Warnow, to = 'daily', method = 'step')
  test.mom.daily$Trave <- interpolate.river.mom(test.mom.monthly$Trave, to = 'daily', method = 'step')
  
  # get netCDF file variable mapping
  varmapping.hbm <- get.varmapping.hbm.netcdf()
  
  # get grid info
  grid_info <- get.infos.grids.hbm.basic()
  
  # get river infos
  file <- 'files/river_list.dat'
  riverInfos <- read.infos.rivers(file, grid_info)
  
  # write new namelist
  write.river.netCDF(c('Warnow', 'Trave'), riverInfos, test.mom.daily, varmapping.hbm, dOt, 'files', 'out_dir', 2012, month = 1, day = 1:5, overwrite=FALSE)

 input vals: mol/kg
 output vals: mmol/m3
 output = input * 20^6
}
\seealso{
write.river.newNML, write.river.newInflow
}
\author{
Daniel Neumann, daniel.neumann@io-warnemuende.de
}
