% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.river.newInflow.R
\name{write.river.newInflow}
\alias{write.river.newInflow}
\title{Creates a new riverinflow* file for the HBM model}
\usage{
write.river.newInflow(riverInfo, riverData, dOt, year, month = 0, day = 0,
  overwrite = TRUE, warn = TRUE)
}
\arguments{
\item{riverInfo}{list with grid information}

\item{riverData}{list with river inflow data}

\item{dOt}{character; directory into which the final file should be written}

\item{year}{integer; year to write out}

\item{month}{integer; month to write out; if month==0 => write all months}

\item{day}{integer; day to write out; if day==0 => write all days}

\item{overwrite}{logical; overwrite 'dOt/filename' it already exists}

\item{warn}{logical; allow/suppress warnings}
}
\value{
file written
}
\description{
A new text file is created. The first row contains " DAILY RIVER RUNOFF" and
the second row "[date in format 'YYYY.MM.DD hh:mm:ss'] [annual inflow 1] [annual inflow 2]"
The 'annual inflow n' has to be printed in the Fortran format 'F8.6' 
(width=8; 6 digits in front of the '.'). One file is created per day.
}
\examples{

  # read a file:
  test.mom.monthly <- read.river.mom('files/GER_Dan_Str_Warnow.dat')
  
  # interpolate from monthly to daily
  test.mom.daily <- interpolate.river.mom(test.mom.monthly, to = 'daily', method = 'step')
  
  # get grid info
  grid_info <- get.infos.grids.hbm.basic()
  
  # get river infos
  file <- 'files/river_list.dat'
  riverInfos <- read.infos.rivers(file, grid_info)
  
  # write new namelist
  write.river.newNML(riverInfos$Warnow, test.mom.daily, 'out_dir', 2012, month = 1, day = 1:5, overwrite=FALSE)
  

 ## example file content
 # DAILY RIVER RUNOFF
 # 2012.07.27 12:00:00     2.1
 
}
\seealso{
write.river.append2Inflow, write.river.newNML, write.river.netCDF
}
\author{
Daniel Neumann, daniel.neumann@io-warnemuende.de
}
