% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.river.newNML.R
\name{write.river.newNML}
\alias{write.river.newNML}
\title{Creates a new grid data namelist file for the HBM model}
\usage{
write.river.newNML(riverInfo, riverData, dOt, year, overwrite = TRUE,
  warn = TRUE)
}
\arguments{
\item{riverInfo}{list with grid information}

\item{riverData}{list with river inflow data}

\item{dOt}{character; directory into which the final file should be written}

\item{year}{integer; year to write out}

\item{overwrite}{logical; overwrite 'dOt/filename' it already exists}

\item{warn}{logical; allow/suppress warnings}
}
\value{
file written
}
\description{
A new fortran namelist file is created. It contains the two nameslists
'DIMENSIONS' and 'RIVERS'. 'RIVERS', contains the variables 'KRQI' and 
'RWQI'. 'KRQI' contains the x- and y-coordinates of the grid cells, in 
which the rivers enter the sea. 'RWQI' contains the actual inflow data 
of each river.
}
\examples{

  # read a file:
  test.mom.monthly <- read.river.mom('files/GER_Dan_Str_Warnow.dat')
  
  # calculate annual means from monthly data
  test.mom.annual <- mean.river.mom(test.mom.monthly, to = 'anual')
  
  # get grid info
  grid_info <- get.infos.grids.hbm.basic()
  
  # get river infos
  file <- 'files/river_list.dat'
  riverInfos <- read.infos.rivers(file, grid_info)
  
  # write new namelist
  write.river.newNML(riverInfos$Warnow, test.mom.annual, 'out_dir', 2012, overwrite=FALSE)
  

  ## example content of an output file:
  # &DIMENSIONS
  # EW=0, NS=0, LAYERS=0, NZBND=0, NUBND=0, NVBND=0, NRIVERS=1, NUDAMS=0, NVDAMS=0, NWEIRS=0
  # /
  # &RIVERS
  # KRQI=0, 0, 253, 289, RWQI=0.000000000000000000E+00, 3.53739587431693980
  # /

&DIMENSIONS
EW=630, NS=387, LAYERS=25, NZBND=858, NUBND=0, NVBND=0, NRIVERS=8, NUDAMS=0, NVDAMS=0, NWEIRS=0
/
&RIVERS
KRQI=0, 0, 341, 606, 376, 82, 386, 169, 364, 302, 263, 193, 84, 579, 26, 477, 65, 159, RWQI=0.000000000000000000E+00, 522.000000000000000, 100.000000000000000, 450.000000000000000, 90
0.000000000000000, 37.0000000000000000, 48.0000000000000000, 21.0000000000000000, 125.000000000000000
/

TODOs:
 - write NRIVERS in DIMENSIONS namelist
 - write coordinates (grid cell indices!) and inflow to RIVERS:
      --Y,X,VAL
      --MONTHLY (m3/s)
}
\seealso{
write.river.netCDF, write.river.newInflow, write.river.append2NML
}
\author{
Daniel Neumann, daniel.neumann@io-warnemuende.de
}
